removed duplicate files
