+++ app/src/main/java/cmput301f17t13/com/catisadog/models/followrequest/FollowRequestRepository.java
@@ -1,13 +1,13 @@
 package cmput301f17t13.com.catisadog.models.followrequest;
 
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
+import cmput301f17t13.com.catisadog.utils.data.OnResultListener;
 import cmput301f17t13.com.catisadog.utils.data.Repository;
-
-/**
- * Created by kevin on 2017-12-02.
- */
 
 public class FollowRequestRepository implements Repository<FollowRequest> {
 
@@ -33,5 +33,21 @@ public class FollowRequestRepository implements Repository<FollowRequest> {
     @Override
     public void delete(String key) {
         mFollowRequestRef.child(key).removeValue();
+    }
+
+    @Override
+    public void get(String key, final OnResultListener<FollowRequest> resultListener) {
+        mFollowRequestRef.child(key).addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                FollowRequest request = dataSnapshot.getValue(FollowRequest.class);
+                resultListener.onResult(request);
+            }
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+
+            }
+        });
     }
 }
